Modified the PropagationControl object to assign a default value for the azimuth of fracture set zero which will be used if the load tensors for all deformation episodes are isotropic, and modified main() (for the standalone version) and ExecuteSimple() (for the Petrel plugin) to set this to the specified minimum horizontal strain azimuth for the first deformation episode. This prevents the fracture orienation defaulting to E-W if another minimum horizontal strain orientation is specified but the strain load is isotropic (B90).
Also added DFM_Plotter_6set_Centrelines.xlsm to data analysis spreadsheet collection.

+++ DFMGenerator_SharedCode/GeometricObjects.cs
@@ -1577,7 +1577,7 @@ namespace DFMGenerator_SharedCode
             double numerator = 2 * components[Tensor2SComponents.XY];
             double denominator = components[Tensor2SComponents.YY] - components[Tensor2SComponents.XX];
 
-            if ((numerator == 0) && (denominator == 0))
+            if (((float)numerator == 0f) && ((float)denominator == 0f))
                 return double.NaN;
             else
                 return (Math.PI + Math.Atan2(numerator, denominator)) / 2;
